+++ src/features/image-area-selector/stores/selection/index.ts
